text-minimessage: Move all internal classes to an internal package
Hide the serializer SPI and internal classes from Javadoc.

Closes GH-530
