+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/internal/package-info.java
@@ -22,7 +22,10 @@
  * SOFTWARE.
  */
 /**
- * Token consumers.
+ * Internal API, not for public use.
+ *
+ * <p>The contents of this package may change at any time, and may not be exported when adventure is present on the module path.
+ * Some elements may be promoted to public API based on user interest.</p>
  */
 @org.jetbrains.annotations.ApiStatus.Internal
-package net.kyori.adventure.text.minimessage.parser.match;
+package net.kyori.adventure.text.minimessage.internal;
