+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/internal/parser/match/TokenListProducingMatchedTokenConsumer.java
@@ -21,62 +21,46 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.parser.match;
+package net.kyori.adventure.text.minimessage.internal.parser.match;
 
-import net.kyori.adventure.text.minimessage.parser.TokenType;
-import org.jetbrains.annotations.MustBeInvokedByOverriders;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import net.kyori.adventure.text.minimessage.internal.parser.Token;
+import net.kyori.adventure.text.minimessage.internal.parser.TokenType;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.UnknownNullability;
 
 /**
- * A consumer of a region of a string that was identified as a token.
+ * A matched token consumer that produces a list of matched tokens.
  *
- * @param <T> the return result
  * @since 4.10.0
  */
-public abstract class MatchedTokenConsumer<T> {
-  protected final String input;
-
-  private int lastIndex = -1;
+public final class TokenListProducingMatchedTokenConsumer extends MatchedTokenConsumer<List<Token>> {
+  private List<Token> result = null;
 
   /**
-   * Creates a new matched token consumer.
+   * Creates a new token list producing matched token consumer.
    *
    * @param input the input
    * @since 4.10.0
    */
-  public MatchedTokenConsumer(final @NotNull String input) {
-    this.input = input;
+  public TokenListProducingMatchedTokenConsumer(final @NotNull String input) {
+    super(input);
   }
 
-  /**
-   * Accepts a matched token.
-   *
-   * @param start     the start of the token
-   * @param end       the end of the token
-   * @param tokenType the type of the token
-   * @since 4.10.0
-   */
-  @MustBeInvokedByOverriders
+  @Override
   public void accept(final int start, final int end, final @NotNull TokenType tokenType) {
-    this.lastIndex = end;
+    super.accept(start, end, tokenType);
+
+    if (this.result == null) {
+      this.result = new ArrayList<>();
+    }
+
+    this.result.add(new Token(start, end, tokenType));
   }
 
-  /**
-   * Gets the result of this consumer, if any.
-   *
-   * @return the result
-   * @since 4.10.0
-   */
-  public abstract @UnknownNullability T result();
-
-  /**
-   * The last accepted end index, or {@code -1} if no match has been accepted.
-   *
-   * @return the last accepted end index
-   * @since 4.10.0
-   */
-  public final int lastEndIndex() {
-    return this.lastIndex;
+  @Override
+  public @NotNull List<Token> result() {
+    return this.result == null ? Collections.emptyList() : this.result;
   }
 }
