+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/internal/parser/node/ValueNode.java
@@ -21,10 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.parser.node;
+package net.kyori.adventure.text.minimessage.internal.parser.node;
 
 import java.util.Objects;
-import net.kyori.adventure.text.minimessage.parser.Token;
+import net.kyori.adventure.text.minimessage.internal.parser.Token;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
