+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/internal/serializer/ComponentClaimingResolverImpl.java
@@ -21,10 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.serializer;
+package net.kyori.adventure.text.minimessage.internal.serializer;
 
 import java.util.Set;
 import java.util.function.BiFunction;
+import java.util.function.Function;
+import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.Context;
 import net.kyori.adventure.text.minimessage.ParsingException;
 import net.kyori.adventure.text.minimessage.tag.Tag;
@@ -33,15 +35,15 @@ import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-final class StyleClaimingResolverImpl implements TagResolver, SerializableResolver.Single {
+class ComponentClaimingResolverImpl implements TagResolver, SerializableResolver.Single {
   private final @NotNull Set<String> names;
   private final @NotNull BiFunction<ArgumentQueue, Context, Tag> handler;
-  private final @NotNull StyleClaim<?> styleClaim;
+  private final @NotNull Function<Component, @Nullable Emitable> componentClaim;
 
-  StyleClaimingResolverImpl(@NotNull final Set<String> names, @NotNull final BiFunction<ArgumentQueue, Context, Tag> handler, @NotNull final StyleClaim<?> styleClaim) {
+  ComponentClaimingResolverImpl(final Set<String> names, final BiFunction<ArgumentQueue, Context, Tag> handler, final Function<Component, @Nullable Emitable> componentClaim) {
     this.names = names;
     this.handler = handler;
-    this.styleClaim = styleClaim;
+    this.componentClaim = componentClaim;
   }
 
   @Override
@@ -57,7 +59,7 @@ final class StyleClaimingResolverImpl implements TagResolver, SerializableResolv
   }
 
   @Override
-  public @Nullable StyleClaim<?> claimStyle() {
-    return this.styleClaim;
+  public @Nullable Emitable claimComponent(final @NotNull Component component) {
+    return this.componentClaim.apply(component);
   }
 }
