+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/internal/serializer/QuotingOverride.java
@@ -21,46 +21,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.parser.match;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import net.kyori.adventure.text.minimessage.parser.Token;
-import net.kyori.adventure.text.minimessage.parser.TokenType;
-import org.jetbrains.annotations.NotNull;
+package net.kyori.adventure.text.minimessage.internal.serializer;
 
 /**
- * A matched token consumer that produces a list of matched tokens.
+ * The preferred style of argument quoting for a specific argument.
+ *
+ * <p>This will allow overriding the style preferences for specific arguments.</p>
  *
  * @since 4.10.0
  */
-public final class TokenListProducingMatchedTokenConsumer extends MatchedTokenConsumer<List<Token>> {
-  private List<Token> result = null;
-
+public enum QuotingOverride {
   /**
-   * Creates a new token list producing matched token consumer.
+   * Request that this argument is included unquoted.
    *
-   * @param input the input
+   * <p>As no escapes can be included in an unquoted argument, this request may be ignored depending on the content of the argument.</p>
+   *
    * @since 4.10.0
    */
-  public TokenListProducingMatchedTokenConsumer(final @NotNull String input) {
-    super(input);
-  }
-
-  @Override
-  public void accept(final int start, final int end, final @NotNull TokenType tokenType) {
-    super.accept(start, end, tokenType);
-
-    if (this.result == null) {
-      this.result = new ArrayList<>();
-    }
-
-    this.result.add(new Token(start, end, tokenType));
-  }
-
-  @Override
-  public @NotNull List<Token> result() {
-    return this.result == null ? Collections.emptyList() : this.result;
-  }
+  UNQUOTED,
+  /**
+   * Request that this argument is included quoted.
+   *
+   * @since 4.10.0
+   */
+  QUOTED;
 }
