+++ text-minimessage/src/main/java/net/kyori/adventure/text/minimessage/internal/serializer/package-info.java
@@ -21,33 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.minimessage.parser.node;
-
-import net.kyori.adventure.text.minimessage.tree.Node;
-import org.jetbrains.annotations.NotNull;
-
 /**
- * Represents the root node of a tree.
+ * Serializer for the MiniMessage language.
+ *
+ * <p>This is currently internal API, likely to change at any time. It may or may not ever become stable.</p>
  *
  * @since 4.10.0
  */
-public final class RootNode extends ElementNode implements Node.Root {
-  private final String beforePreprocessing;
-
-  /**
-   * Creates a new root node.
-   *
-   * @param sourceMessage the source message
-   * @param beforePreprocessing the source message before handling preProcess tags
-   * @since 4.10.0
-   */
-  public RootNode(final @NotNull String sourceMessage, final @NotNull String beforePreprocessing) {
-    super(null, null, sourceMessage);
-    this.beforePreprocessing = beforePreprocessing;
-  }
-
-  @Override
-  public @NotNull String input() {
-    return this.beforePreprocessing;
-  }
-}
+@org.jetbrains.annotations.ApiStatus.Internal
+package net.kyori.adventure.text.minimessage.internal.serializer;
